Move back to single catalog integration tests
