+++ kafka-connect-runtime/src/test/java/io/tabular/iceberg/connect/IntegrationDynamicTableTest.java
@@ -36,7 +36,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public abstract class AbstractIntegrationDynamicTableTest extends IntegrationTestBase {
+public class IntegrationDynamicTableTest extends IntegrationTestBase {
 
   private static final String TEST_DB = "test";
   private static final String TEST_TABLE1 = "tbl1";
@@ -52,7 +52,7 @@ public abstract class AbstractIntegrationDynamicTableTest extends IntegrationTes
 
   @AfterEach
   public void after() {
-    context.stopKafkaConnector(connectorName);
+    context.stopConnector(connectorName);
     deleteTopic(testTopic);
     catalog.dropTable(TableIdentifier.of(TEST_DB, TEST_TABLE1));
     catalog.dropTable(TableIdentifier.of(TEST_DB, TEST_TABLE2));
@@ -98,13 +98,13 @@ public abstract class AbstractIntegrationDynamicTableTest extends IntegrationTes
             .config("iceberg.control.commit.interval-ms", 1000)
             .config("iceberg.control.commit.timeout-ms", Integer.MAX_VALUE)
             .config("iceberg.kafka.auto.offset.reset", "earliest");
-    connectorCatalogProperties().forEach(connectorConfig::config);
+    context.connectorCatalogProperties().forEach(connectorConfig::config);
 
     if (branch != null) {
       connectorConfig.config("iceberg.tables.default-commit-branch", branch);
     }
 
-    context.startKafkaConnector(connectorConfig);
+    context.startConnector(connectorConfig);
 
     TestEvent event1 =
         new TestEvent(1, "type1", System.currentTimeMillis(), TEST_DB + "." + TEST_TABLE1);
